requirements.txt
# Windowing and drawing
pysdl2==0.9.5
